fix copy-paste of sessionInfo - added common internal/safe.SessionInfo helper
