+++ internal/query/client.go
@@ -17,6 +17,7 @@ import (
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/query/gtrace"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/query/options"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/query/result"
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/safe"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/stack"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/tx"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/types"
@@ -826,20 +827,6 @@ func newWithQueryServiceClient(ctx context.Context,
 	return c, nil
 }
 
-func sessionInfo[PT pool.ItemConstraint[T], T any](item PT) trace.SessionInfo {
-	// PT is *T; compare with nil before interface conversion to avoid typed nil.
-	if item == nil {
-		return nil
-	}
-
-	s, ok := any(item).(trace.SessionInfo)
-	if !ok {
-		return nil
-	}
-
-	return s
-}
-
 func poolTrace(t *trace.Query) *pool.Trace[*Session, Session] {
 	return &pool.Trace[*Session, Session]{
 		OnNew: func(ctx *context.Context, call stack.Caller) func(limit int) {
@@ -871,7 +858,7 @@ func poolTrace(t *trace.Query) *pool.Trace[*Session, Session] {
 			}
 		},
 		OnPut: func(ctx *context.Context, call stack.Caller, item *Session) func(err error) {
-			onDone := gtrace.QueryOnPoolPut(t, ctx, call, sessionInfo(item))
+			onDone := gtrace.QueryOnPoolPut(t, ctx, call, safe.SessionInfo(item))
 
 			return func(err error) {
 				onDone(err)
@@ -886,7 +873,7 @@ func poolTrace(t *trace.Query) *pool.Trace[*Session, Session] {
 			onDone := gtrace.QueryOnPoolGet(t, ctx, call)
 
 			return func(session *Session, hint *trace.NodeHintInfo, attempts int, err error) {
-				onDone(sessionInfo(session), attempts, hint, err)
+				onDone(safe.SessionInfo(session), attempts, hint, err)
 			}
 		},
 		OnChange: func(stats pool.Stats) {
