+++ internal/query/client_test.go
@@ -1999,43 +1999,3 @@ func TestQueryScript(t *testing.T) {
 		})
 	})
 }
-
-type sessionInfoItem struct {
-	id string
-}
-
-func (s *sessionInfoItem) ID() string                  { return s.id }
-func (s *sessionInfoItem) NodeID() uint32              { return 1 }
-func (s *sessionInfoItem) Status() string              { return "ready" }
-func (s *sessionInfoItem) Close(context.Context) error { return nil }
-func (s *sessionInfoItem) IsAlive() bool               { return true }
-
-func TestSessionInfo(t *testing.T) {
-	t.Parallel()
-
-	t.Run("NilPointer", func(t *testing.T) {
-		t.Parallel()
-
-		var item *sessionInfoItem
-
-		session := sessionInfo(item)
-		require.Nil(t, session)
-	})
-	t.Run("TypedNilIsUntypedNil", func(t *testing.T) {
-		t.Parallel()
-
-		session := sessionInfo((*sessionInfoItem)(nil))
-
-		var untyped trace.SessionInfo
-		require.Equal(t, untyped, session)
-	})
-	t.Run("NonNil", func(t *testing.T) {
-		t.Parallel()
-
-		item := &sessionInfoItem{id: "test"}
-		session := sessionInfo(item)
-
-		require.NotNil(t, session)
-		require.Equal(t, "test", session.ID())
-	})
-}
